+++ packages/frontend/src/utils/project/contracts-and-permissions/getContractsSection.ts
@@ -1,4 +1,5 @@
 import type {
+  Project,
   ProjectContract,
   ProjectContracts,
   ProjectEscrow,
@@ -15,6 +16,7 @@ import type { ContractsSectionProps } from '../../../components/projects/section
 import { toTechnologyRisk } from '../risk-summary/toTechnologyRisk'
 import type { ContractUtils } from './getContractUtils'
 import { getPastUpgradesData } from './getPastUpgradesData'
+import { getZkProgramHashes } from './getZkProgramHashes'
 import { toVerificationStatus } from './toVerificationStatus'
 
 type ProjectParams = {
@@ -35,6 +37,8 @@ export function getContractsSection(
   projectParams: ProjectParams,
   contractUtils: ContractUtils,
   projectsChangeReport: ProjectsChangeReport,
+  zkCatalogProjects: Project<'zkCatalogInfo'>[],
+  allProjects: Project<'contracts'>[],
 ): ContractsSection | undefined {
   if (!projectParams.contracts) {
     return undefined
@@ -93,6 +97,11 @@ export function getContractsSection(
       projectParams.architectureImage ?? projectParams.slug,
     ),
     isUnderReview: projectParams.isUnderReview,
+    zkProgramHashes: getZkProgramHashes(
+      projectParams.contracts.zkProgramHashes,
+      zkCatalogProjects,
+      allProjects,
+    ),
   }
 }
